+++ src/CounterInterface.php
@@ -17,8 +17,6 @@ interface CounterInterface
 
     /**
      * Counts one request as done and returns object containing current counter state.
-     *
-     * @return CounterState
      */
     public function incrementAndGetState(): CounterState;
 }