+++ CHANGELOG.md
@@ -1,9 +1,8 @@
-# RateLimiter Change Log
-
+# Yii Rate Limiter Middleware Change Log
 
 ## 1.0.1 under development
 
-- no changes in this release.
+- Bug #14: Throw exception on call `getCacheKey()` in counter without the specified ID (vjik)
 
 ## 1.0.0 June 07, 2021